requirements.txt
Flask==3.0.0
google-api-python-client==2.108.0
google-auth==2.23.4
python-docx==1.1.0
